Incomplete integration of the prior sprint
I have copied the Grid.ipynb into grid.py, updating it so that the graph function will now output a string that can be read into an image by the browser.  The image is NOT saved to the device through this method, as I believe it will be quite a hassle.  A portion of the code was copied from https://gitlab.com/-/snippets/1924163 for that part.  I also updated app.py to, when logged in, display the grid image.  The current setup will display a square, however it should be adjusted based on the grid parameters later.

Start.ps1 was adjusted to also install matplotlib, it will probably need to be adjusted later during this sprint to install other libraries.

I have yet to integrate the text-based representation of the maze or the adjustable size via inputs or the connectivity features.