Integrated features and expanded tests
Updated app.py to display the grid whether the user is signed in or not, should the user not be signed in they will currently be able to see the text version of the grid.

Updated main.py to call test.py whenever DEBUG is true so testing can be facilitated more efficiently.

Updated index.html to display an image and the text version of the image scaled to be similar to the image.

Created test.py which is majorly a duplication of Hayden's test work in the prior commit, but also expands on it by adding a few more tests and makes it easier to tell when a test passes or fails.  There are still more tests to be created.

Updated grid.py to not rely on numpy, also did a major revamp on how Grid works.  Grid now is initialized using "sampleGrid = Grid( x : int, y : int, seed : int)", and it automatically creates a grid that can be accessed using "sampleGrid.grid".  Grid also has generateRooms as a method, which takes in an integer for the amount of rooms and another integer for the maximum room size.  The method uses the Room class, which was also adjusted to suit the new grid's shape.  Within the Grid class, graph and image were separated as they are used in different types of applications.  The Grid class also has the text method, which was duplicated from Mara's ASCIIGrid.ipynb then altered.

Inputs have not been integrated due to lack of knowledge regarding the process.