Added the general idea for hover descriptions
Updated the html files to loop through a list with lists of the coords and descriptions for each room of the maze.  The hover not only pops up a toolTip, but also changes the cursor icon and upon click prints out the description for testing purposes currently.

App.py now has a sample function, grab_map, that returns a sample output to populate the html files with the image maps for hovering.

I am uncertain the best approach to convert grid location to image location, but it may involve looking through matplotlib's source code.  So far I have found out about BBox in transforms.py and FigureCanvasBase in backend_bases.py.  So it MAY be in there, or figure.py.