Setup tests for grid to image location
In app.py, adjusted the hover location to be the top-left corner of the grid.  It will be adjusted later once room description has been integrated.

In grid.py, added a method to Grid called toImageLocation, that converts the provided rectangle on the grid into a rectangle within the output image when using the Grid.image method.

In test.py, added tests for the Grid.toImageLocation method and finished the tests for the colorToString.