Cleaned up
Updated the branch to be consistent w/ the main branch, adding Zak and Coby's changes and refixing the minor sqllite3 error.

For app.py, changed the grab_map function to be more efficient space and speed wise, however now room locations are stored within the grid, which negates the memory saved.  The output html files are also much smaller, since prior they created a 1 by 1 area for each pixel, which made the html files extremely large.  The names of the rooms were also based on the color instead of the actual room type, thus they were altered to be more understandable.

For the html files, changed onmoveover to onmovemove to make it smoother since they no longer operate on a pixel basis.  Removed the title css and changed it to alt.

For grid.py Grid class, I added a way to remember the rooms and their respective colors, which is in part used by the map_grid function in app, but was primarily created to assist in speed when generating the maze path, as we can tell whether we are within the rooms or not.  I also added a rudimentary description of how I think generatePath will operate, so when customization of the maze complexity is undergone, it is easier.

TODO: Finish the generation of paths, and clean up the ever expanding grid.py  and app.py and test.py files, using a modular approach.