Integrated Maze path and connectivity
Coby's ensure connectivity was modified to randomly chose the first direction to look at, since it always started looking in a set one, which in some cases didn't open up closed off locations.  This approach reduces the chance for a room to be closed off from the rest of the rooms.